Move `::lock()` to `ModelWithContent`
