+++ tests/Panel/ModelTest.php
@@ -5,7 +5,6 @@ namespace Kirby\Panel;
 use Kirby\Cms\App;
 use Kirby\Cms\File as ModelFile;
 use Kirby\Cms\Site as ModelSite;
-use Kirby\Content\Lock;
 use Kirby\Filesystem\Asset;
 use Kirby\Filesystem\Dir;
 use Kirby\TestCase;
@@ -382,20 +381,6 @@ class ModelTest extends TestCase
 	{
 		$this->assertIsString(Model::imagePlaceholder());
 		$this->assertStringStartsWith('data:image/gif;base64,', Model::imagePlaceholder());
-	}
-
-	/**
-	 * @covers ::lock
-	 */
-	public function testLock()
-	{
-		$site = new ModelSite();
-		$lock = $site->panel()->lock();
-
-		$this->assertInstanceOf(Lock::class, $lock);
-		$this->assertFalse($lock->isLocked());
-		$this->assertNull($lock->modified());
-		$this->assertNull($lock->user());
 	}
 
 	/**
